requirements.txt
Flask==2.1.0
Pillow==9.0.1
